Move GRIBCode code into a separate submodule.
